requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Face recognition and image processing - Windows compatible
opencv-python==4.8.1.78
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
scikit-learn==1.3.2
mediapipe==0.10.8

# Database
sqlalchemy>=2.0.35
psycopg2-binary==2.9.11  # PostgreSQL adapter

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
